+++ packages/react-three-cannon-examples/src/colors.ts
@@ -1,0 +1,2 @@
+// From `nice-color-palettes`, this one is index 17
+export default ['#99b898', '#fecea8', '#ff847c', '#e84a5f', '#2a363b'] as const